remote submission refactor initial, untested
