+++ src/alomancy/remote_submission/__init__.py
@@ -1,0 +1,15 @@
+from alomancy.remote_submission.executor import RemoteJobExecutor
+from alomancy.remote_submission.submitters import (
+    all_maces_remote_submitter,
+    committee_remote_submitter,
+    md_remote_submitter,
+    qe_remote_submitter,
+)
+
+__all__ = [
+    "RemoteJobExecutor",
+    "all_maces_remote_submitter",
+    "committee_remote_submitter",
+    "md_remote_submitter",
+    "qe_remote_submitter",
+]
