+++ src/alomancy/remote_submission/executor.py
@@ -8,6 +8,8 @@ from expyre.func import ExPyRe
 from alomancy.configs.remote_info import RemoteInfo
 
 logger = logging.getLogger(__name__)
+
+
 class RemoteJobExecutor:
     """
     General-purpose remote job submission utility.
@@ -17,14 +19,6 @@ class RemoteJobExecutor:
     """
 
     def __init__(self, remote_info: RemoteInfo):
-        """
-        Initialize the remote executor.
-
-        Parameters
-        ----------
-        remote_info : RemoteInfo
-            Configuration for remote execution
-        """
         self.remote_info = remote_info
         self.jobs = []
 
@@ -37,45 +31,17 @@ class RemoteJobExecutor:
         job_name: Optional[str] | None = None,
         **expyre_kwargs,
     ) -> ExPyRe:
-        """
-        Submit a single job to remote execution.
-
-        Parameters
-        ----------
-        function : Callable
-            The function to execute remotely
-        function_kwargs : Dict[str, Any]
-            Keyword arguments to pass to the function
-        input_files : List[Union[str, Path]], optional
-            Files to transfer to remote system
-        output_files : List[Union[str, Path]], optional
-            Files to transfer back from remote system
-        job_name : str, optional
-            Name for this specific job (overrides remote_info.job_name)
-        **expyre_kwargs
-            Additional ExPyRe-specific arguments
-
-        Returns
-        -------
-        ExPyRe
-            The ExPyRe job object
-        """
-
         if input_files is None:
             input_files = []
         if output_files is None:
             output_files = []
         if job_name is None:
             job_name = self.remote_info.job_name
-        # Convert paths to strings
         input_files = [str(f) for f in (input_files or [])]
         output_files = [str(f) for f in (output_files or [])]
 
-        # Use provided files or fall back to remote_info defaults
         final_input_files = input_files or self.remote_info.input_files
-        final_output_files = output_files or getattr(
-            self.remote_info, "output_files", []
-        )
+        final_output_files = output_files or getattr(self.remote_info, "output_files", [])
 
         job = ExPyRe(
             name=job_name or self.remote_info.job_name,
@@ -100,31 +66,6 @@ class RemoteJobExecutor:
         common_output_pattern: Optional[str] | None = None,
         job_name_pattern: Optional[str] | None = None,
     ) -> list[ExPyRe]:
-        """
-        Submit multiple similar jobs with different parameters.
-
-        Parameters
-        ----------
-        function : Callable
-            The function to execute remotely
-        job_configs : List[Dict[str, Any]]
-            List of job configurations, each containing:
-            - function_kwargs: Dict of kwargs for the function
-            - input_files: Optional list of input files (in addition to common)
-            - output_files: Optional list of output files
-            - job_name: Optional specific job name
-        common_input_files : List[Union[str, Path]], optional
-            Input files common to all jobs
-        common_output_pattern : str, optional
-            Pattern for output files, use {job_id} for job index
-        job_name_pattern : str, optional
-            Pattern for job names, use {job_id} for job index
-
-        Returns
-        -------
-        List[ExPyRe]
-            List of ExPyRe job objects
-        """
         if common_input_files is None:
             common_input_files = []
         if job_name_pattern is None:
@@ -134,17 +75,15 @@ class RemoteJobExecutor:
         common_input_files = common_input_files or []
 
         for i, config in enumerate(job_configs):
-            # Prepare input files
             job_input_files = list(common_input_files)
             if "input_files" in config:
                 job_input_files.extend(config["input_files"])
 
-            # Prepare output files
             job_output_files = config.get("output_files", [])
             if common_output_pattern:
                 job_output_files.append(common_output_pattern.format(job_id=i))
             logger.debug("Job %d output files: %s", i, job_output_files)
-            # Prepare job name
+
             job_name = config.get("job_name")
             if not job_name and job_name_pattern:
                 job_name = job_name_pattern.format(job_id=i)
@@ -161,14 +100,6 @@ class RemoteJobExecutor:
         return jobs
 
     def start_all_jobs(self, **start_kwargs) -> None:
-        """
-        Start all submitted jobs.
-
-        Parameters
-        ----------
-        **start_kwargs
-            Additional arguments for job.start()
-        """
         for job in self.jobs:
             job.start(
                 resources=self.remote_info.resources,
@@ -180,14 +111,6 @@ class RemoteJobExecutor:
             )
 
     def wait_for_all_jobs(self) -> list[Any]:
-        """
-        Wait for all jobs to complete and gather results.
-
-        Returns
-        -------
-        List[Any]
-            List of results from all jobs
-        """
         results = []
 
         for i, job in enumerate(self.jobs):
@@ -212,7 +135,6 @@ class RemoteJobExecutor:
         return results
 
     def cleanup_jobs(self) -> None:
-        """Mark all jobs as processed for cleanup."""
         for job in self.jobs:
             job.mark_processed()
 
@@ -222,75 +144,12 @@ class RemoteJobExecutor:
         job_configs: list[dict[str, Any]],
         **kwargs,
     ) -> list[Any]:
-        """
-        Convenience method to submit, start, and wait for multiple jobs.
-
-        Parameters
-        ----------
-        function : Callable
-            The function to execute remotely
-        job_configs : List[Dict[str, Any]]
-            List of job configurations
-        **kwargs
-            Additional arguments for submit_multiple_jobs
-
-        Returns
-        -------
-        List[Any]
-            Results from all jobs
-        """
         logger.debug("run_and_wait working directory: %s", os.getcwd())
         self.submit_multiple_jobs(function, job_configs, **kwargs)
         self.start_all_jobs()
         self.wait_for_all_jobs()
 
-        # final run of this essential to get results to sync locally
+        # final call essential to sync results locally from the remote
         results = self.wait_for_all_jobs()
         self.cleanup_jobs()
         return results
-
-
-# Convenience functions for backward compatibility
-# def submit_committee_jobs(
-#     remote_info: RemoteInfo,
-#     function: Callable,
-#     function_kwargs: Dict[str, Any],
-#     base_name: str,
-#     size_of_committee: int = 5,
-#     **kwargs
-# ) -> List[Any]:
-#     """
-#     Submit committee of jobs (like MACE ensemble training).
-
-#     This maintains backward compatibility with your original use case.
-#     """
-#     executor = RemoteJobExecutor(remote_info)
-
-#     workdir = Path("results", base_name)
-
-#     # Create job configs for committee
-#     job_configs = []
-#     for i in range(size_of_committee):
-#         job_configs.append({
-#             'function_kwargs': {
-#                 **function_kwargs,
-#                 'seed': function_kwargs.get('seed', 803) + i,  # Different seed per job
-#                 'output_dir': str(workdir / "MACE" / f"fit_{i}")
-#             },
-#             'output_files': [str(workdir / "MACE" / f"fit_{i}")]
-#         })
-
-#     # Common input files
-#     common_input_files = [
-#         "mace_wfl.py",
-#         str(workdir / "train_set.xyz"),
-#         str(workdir / "test_set.xyz"),
-#     ]
-
-#     return executor.run_and_wait(
-#         function=function,
-#         job_configs=job_configs,
-#         common_input_files=common_input_files,
-#         job_name_pattern=f"{remote_info.job_name}_{{job_id}}",
-#         **kwargs
-#     )
